Typescriptify some of the drawer widgets
