+++ packages/core/util/types/index.ts
@@ -122,18 +122,35 @@ export function isSessionModelWithConfigEditing(
   return isSessionModel(thing) && 'editConfiguration' in thing
 }
 
+export interface Widget {
+  type: string
+  id: string
+}
+
 /** abstract interface for a session that manages widgets */
 export interface SessionWithWidgets extends AbstractSessionModel {
-  visibleWidget?: { id: string }
-  widgets: Map<string, unknown>
+  minimized: boolean
+  visibleWidget?: Widget
+  widgets: Map<string, Widget>
+  activeWidgets: Map<string, Widget>
   addWidget(
     typeName: string,
     id: string,
     initialState?: Record<string, unknown>,
     configuration?: { type: string },
-  ): void
+  ): Widget
   showWidget(widget: unknown): void
+  hideWidget(widget: unknown): void
 }
+
+/* only some sessions with widgets use a drawer widget */
+export interface SessionWithDrawerWidgets extends SessionWithWidgets {
+  drawerWidth: number
+  resizeDrawer(arg: number): number
+  minimizeWidgetDrawer(): void
+  showWidgetDrawer: () => void
+}
+
 export function isSessionModelWithWidgets(
   thing: unknown,
 ): thing is SessionWithWidgets {
